# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transparência API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/despesas": {
            "get": {
                "tags": [
                    "/despesas"
                ],
                "summary": "listar despesas",
                "operationId": "ac8e14d119350f34662d57c5912ccca0",
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/despesas"
                ],
                "summary": "Registrar uma despesa",
                "operationId": "04942cfc96912eb8a8639b8e8aa7defa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_objedespe": {
                                        "description": "objectivo da despesa",
                                        "type": "string"
                                    },
                                    "n_codicoord": {
                                        "description": "id do coordenador que criou a despesa",
                                        "type": "integer"
                                    },
                                    "n_valodespe": {
                                        "description": "valores da deespesa",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "c_fontdespe": {
                                        "description": "fonte dos valores",
                                        "type": "integer",
                                        "default": "caixa"
                                    },
                                    "d_dasadespe": {
                                        "description": "data do saque dos valores",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "despesa cadastrado com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/despesas/coord/{idCoordPredio}": {
            "get": {
                "tags": [
                    "/despesas"
                ],
                "summary": "mostrar um despesa",
                "operationId": "f3e17e718fb9f3162e107a97e54e425c",
                "parameters": [
                    {
                        "name": "idCoordPredio",
                        "in": "path",
                        "description": "id do coordenador do predio",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "despesas não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/despesas/{despesa}": {
            "get": {
                "tags": [
                    "/despesas"
                ],
                "summary": "mostrar um despesa",
                "operationId": "e98b11fc5ff3dca2fa35ee8db499c7c2",
                "parameters": [
                    {
                        "name": "despesa",
                        "in": "path",
                        "description": "id do despesa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "despesa não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/despesas"
                ],
                "summary": "apagar uma despesas",
                "operationId": "f6980c1fb7b2523814b8b0e5be5a61da",
                "parameters": [
                    {
                        "name": "despesa",
                        "in": "path",
                        "description": "id do despesa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "despesa deletada com sucesso!"
                    },
                    "404": {
                        "description": "despesas não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dividas": {
            "get": {
                "tags": [
                    "/dividas"
                ],
                "summary": "listar dividas",
                "operationId": "a96c0f55b32afd91945f1d7aa29962e1",
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dividas/apartamento/{apartamento}": {
            "get": {
                "tags": [
                    "/dividas"
                ],
                "summary": "mostrar divida do apartamento",
                "operationId": "00ac21dfd51489764236edc7a65f85d5",
                "parameters": [
                    {
                        "name": "apartamento",
                        "in": "path",
                        "description": "id do apartamento",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "apartamento não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dividas/{divida}": {
            "get": {
                "tags": [
                    "/dividas"
                ],
                "summary": "mostrar divida",
                "operationId": "15c81bb3c492131338f7ef11e2461a0e",
                "parameters": [
                    {
                        "name": "divida",
                        "in": "path",
                        "description": "id do divida",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "divida não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}